requirements.txt
click==8.1.7
matplotlib==3.9.0
pandas==2.2.2
pydantic==2.6.3
pyyaml==6.0.1